Refactored STKConnManager
The class is now able to build the connectivity for blocks
of geom dim lower than the mesh dim. In particular, we can
build a STKConnManager from a 3d bulk/meta data pair, which
is later used to build connectivity for N 2d parts.
The input parts must all be of the same geo dim, and cannot intersect.